Make a wrapper for the pages to save the state and read it. also handling the loading.
