+++ pages/highscores.jsx
@@ -1,11 +1,15 @@
+import React from 'react';
+import LoadStateWrapper from '../src/components/LoadStateWrapper';
 import ControlPanel from '../src/components/ControlPanel';
 
 function Highscores() {
   return (
-    <div>
-      <div>Highscores</div>
-      <ControlPanel navigationValue="/highscores" />
-    </div>
+    <LoadStateWrapper>
+      <div>
+        <div>Highscores</div>
+        <ControlPanel navigationValue="/highscores" />
+      </div>
+    </LoadStateWrapper>
   );
 }
