+++ pages/index.jsx
@@ -1,27 +1,16 @@
-import React, { useEffect, useContext, useState } from 'react';
+import React from 'react';
 import ControlPanel from '../src/components/ControlPanel';
 import Game from '../src/components/Game';
-import GlobalContext from '../src/components/GlobalContext';
+import LoadStateWrapper from '../src/components/LoadStateWrapper';
 
 function Index() {
-  const { settings, dispatch } = useContext(GlobalContext);
-  const [loaded, setLoaded] = useState(false);
-
-  useEffect(() => {
-    setLoaded(true);
-    const state = localStorage.getItem('state');
-    if (state) dispatch({ type: 'LOAD_STATE', payload: JSON.parse(state) });
-  }, []);
-
-  useEffect(() => {
-    if (loaded) localStorage.setItem('state', JSON.stringify(settings));
-  }, [settings]);
-
   return (
-    <div>
-      <Game />
-      <ControlPanel navigationValue="/" />
-    </div>
+    <LoadStateWrapper>
+      <div>
+        <Game />
+        <ControlPanel navigationValue="/" />
+      </div>
+    </LoadStateWrapper>
   );
 }
